requirements-freeze.txt
accelerate==1.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.10.0
appnope==0.1.4
asttokens==3.0.0
attrs==24.3.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
comm==0.2.2
dataclasses-json==0.6.7
datasets==3.1.0
debugpy==1.8.11
decorator==5.1.1
dill==0.3.8
executing==2.1.0
fastapi==0.116.1
filelock==3.16.1
flasgger==0.9.7b2
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.5.0
fsspec==2024.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.27.0
idna==3.10
ipykernel==6.29.5
ipython==8.30.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.76
langchain-text-splitters==0.3.11
langsmith==0.4.28
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mistune==3.0.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
ollama==0.5.4
orjson==3.11.3
packaging==24.2
pandas==2.2.3
parso==0.8.4
peft==0.13.2
pexpect==4.9.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.1
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.22.3
safetensors==0.4.5
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
stack-data==0.6.3
starlette==0.47.3
sympy==1.13.1
tenacity==9.1.2
tokenizers==0.21.0
torch==2.5.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uv==0.8.17
uvicorn==0.35.0
wcwidth==0.2.13
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.18.3
zstandard==0.25.0
